+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Impl HasRawDisplayHandle for windows.
 - Fix MenuItem::set_callback().
 - Add MenuItem::set_shortcut and set_flag.
+- Use bitflags for MenuFlag.
 
 ## [1.3.20] - 2022-11-13
 - Add ColorChooser::set_tuple_rgb by @ConsoleC137.